+++ packages/react-core/src/components/ToggleGroup/ToggleGroup.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/react-styles/css/components/ToggleGroup/toggle-group';
+import { ToggleGroupItem } from './ToggleGroupItem';
 
 export interface ToggleGroupProps extends React.HTMLProps<HTMLDivElement> {
   /** Content rendered inside the toggle group */
@@ -9,6 +10,8 @@ export interface ToggleGroupProps extends React.HTMLProps<HTMLDivElement> {
   className?: string;
   /** Modifies the toggle group to include compact styling. */
   isCompact?: boolean;
+  /** Disable all toggle group items under this component. */
+  areAllGroupsDisabled?: boolean;
   /** Accessible label for the toggle group */
   'aria-label'?: string;
 }
@@ -17,12 +20,15 @@ export const ToggleGroup: React.FunctionComponent<ToggleGroupProps> = ({
   className,
   children,
   isCompact = false,
+  areAllGroupsDisabled = false,
   'aria-label': ariaLabel,
   ...props
 }: ToggleGroupProps) => {
-  const toggleGroupItemList = [] as any[];
-  React.Children.forEach(children, child => {
-    toggleGroupItemList.push(child);
+  const toggleGroupItemList = React.Children.map(children, child => {
+    const childCompName = (child as any).type.name;
+    return childCompName !== ToggleGroupItem.name
+      ? child
+      : React.cloneElement(child as React.ReactElement, areAllGroupsDisabled ? { isDisabled: true } : {});
   });
 
   return (